Rewrote hide_method_args() -> replace_parens().
Much better at handling nested params now. Using a looped regex to
replace parentheses with an unused unicode point. When there are no more
parentheses, it replaces the unicode point with '(...)'.

This would have been inpossible with standard regex's because they are
based on finite automata which can't handle arbitrary levels of nesting.